requirements.txt
xgboost==0.80
numpy==1.14.3
scipy==1.2.1
matplotlib
scikit-learn==0.19.2
cassandra-driver==3.14.0
pandas==0.23.0
repositorytools==4.2.5
category_encoders==1.3.0
spark-sklearn==0.3.0
pyspark==2.4.4
findspark==1.3.0
requests==2.22.0